Scrape race information from procyclingstats.com.

This function is for internal use by cyclingscraper and racedata to retrieve information about stages of the requested race.

Usage

raceinfo(race, year)

Arguments

race

A character race identifier. See below for details.

year

A number corresponding to requested race year with the format: YYYY.

Value

a dataframe containing race information for race in year with the following variables included: stage, stage_id and stage_type.

Details

The race parameter accepts the following inputs:

"tdf" (Tour de France)

"giro" (Giro d'Italia)

"vuelta" (La Vuelta ciclista a España)

"dauphine" (Critérium du Dauphiné)

"suisse" (Tour de Suisse)

"tirreno" (Tirreno-Adriatico)

"parisnice" (Paris-Nice)

"romandie" (Tour de Romandie)

"catalunya" (Volta Ciclista a Catalunya)

"pologne" (Tour de Pologne)

"basque" (Itzulia Basque Country)

Examples

raceinfo("tdf", 2022)
#> # A tibble: 21 × 3
#>    stage                                                    stage_id stage_type
#>    <chr>                                                    <chr>    <chr>     
#>  1 Stage 1 (ITT) | Copenhagen - Copenhagen                  1        ITT       
#>  2 Stage 2 | Roskilde - Nyborg                              2        Road race 
#>  3 Stage 3 | Vejle - Sønderborg                             3        Road race 
#>  4 Stage 4 | Dunkerque - Calais                             4        Road race 
#>  5 Stage 5 | Lille - Wallers-Arenberg                       5        Road race 
#>  6 Stage 6 | Binche - Longwy                                6        Road race 
#>  7 Stage 7 | Tomblaine - La Super Planche des Belles Filles 7        Road race 
#>  8 Stage 8 | Dole - Lausanne                                8        Road race 
#>  9 Stage 9 | Aigle - Châtel les portes du Soleil            9        Road race 
#> 10 Stage 10 | Morzine - Megève                              10       Road race 
#> # ℹ 11 more rows