List of years where races were completed.

This function is for internal use by cyclingscraper and checkrace to ensure that HTML requests are only passed for years were a given race exists.

Usage

raceyears()

Value

A list containing race years for each race

Examples

raceyears()
#> $tdf
#>   [1] 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912 1913 1914 1919 1920 1921
#>  [16] 1922 1923 1924 1925 1926 1927 1928 1929 1930 1931 1932 1933 1934 1935 1936
#>  [31] 1937 1938 1939 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958
#>  [46] 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973
#>  [61] 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988
#>  [76] 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003
#>  [91] 2004 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018
#> [106] 2019 2020 2021 2022 2023 2024
#> 
#> $giro
#>   [1] 1909 1910 1911 1912 1913 1914 1919 1920 1921 1922 1923 1924 1925 1926 1927
#>  [16] 1928 1929 1930 1931 1932 1933 1934 1935 1936 1937 1938 1939 1940 1946 1947
#>  [31] 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962
#>  [46] 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977
#>  [61] 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992
#>  [76] 1993 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007
#>  [91] 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018 2019 2020 2021 2022
#> [106] 2023 2024
#> 
#> $vuelta
#>  [1] 1935 1936 1941 1942 1945 1946 1947 1948 1950 1955 1956 1957 1958 1959 1960
#> [16] 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975
#> [31] 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#> [46] 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005
#> [61] 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018 2019 2020
#> [76] 2021 2022 2023 2024
#> 
#> $dauphine
#>  [1] 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961
#> [16] 1962 1963 1964 1965 1966 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978
#> [31] 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992 1993
#> [46] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [61] 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018 2019 2020 2021 2022 2023
#> [76] 2024
#> 
#> $suisse
#>  [1] 1933 1934 1935 1936 1937 1938 1939 1941 1942 1946 1947 1948 1949 1950 1951
#> [16] 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966
#> [31] 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981
#> [46] 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996
#> [61] 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011
#> [76] 2012 2013 2014 2015 2016 2017 2018 2019 2021 2022 2023 2024
#> 
#> $tirreno
#>  [1] 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980
#> [16] 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995
#> [31] 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010
#> [46] 2011 2012 2013 2014 2015 2016 2017 2018 2019 2020 2021 2022 2023 2024
#> 
#> $parisnice
#>  [1] 1933 1934 1935 1936 1937 1938 1939 1946 1951 1952 1953 1954 1955 1956 1957
#> [16] 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972
#> [31] 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987
#> [46] 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000 2001 2002
#> [61] 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017
#> [76] 2018 2019 2020 2021 2022 2023 2024
#> 
#> $romandie
#>  [1] 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961
#> [16] 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976
#> [31] 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991
#> [46] 1992 1993 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006
#> [61] 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018 2019 2021 2022
#> [76] 2023 2024
#> 
#> $catalunya
#>   [1] 1911 1912 1913 1920 1923 1924 1925 1926 1927 1928 1929 1930 1931 1932 1933
#>  [16] 1934 1935 1936 1939 1940 1941 1942 1943 1944 1945 1946 1947 1948 1949 1950
#>  [31] 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962 1963 1964 1965
#>  [46] 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980
#>  [61] 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995
#>  [76] 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010
#>  [91] 2011 2012 2013 2014 2015 2016 2017 2018 2019 2021 2022 2023 2024
#> 
#> $pologne
#>  [1] 1928 1929 1937 1939 1947 1948 1949 1952 1953 1954 1955 1956 1957 1958 1959
#> [16] 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974
#> [31] 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989
#> [46] 1990 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004
#> [61] 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018 2019
#> [76] 2020 2021 2022 2023 2024
#> 
#> $basque
#>  [1] 1924 1925 1926 1927 1928 1929 1930 1935 1969 1970 1971 1972 1973 1974 1975
#> [16] 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#> [31] 1991 1992 1993 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005
#> [46] 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018 2019 2021
#> [61] 2022 2023 2024
#>